+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/ui/auth/LoginWebViewActivity.kt
@@ -12,7 +12,7 @@ import androidx.lifecycle.ViewModelProvider
 import com.takseha.data.dto.auth.login.LoginRequest
 import com.takseha.presentation.R
 import com.takseha.presentation.databinding.ActivityLoginWebviewBinding
-import com.takseha.presentation.viewmodel.LoginWebViewViewModel
+import com.takseha.presentation.viewmodel.auth.LoginWebViewViewModel
 
 class LoginWebViewActivity : AppCompatActivity() {
     private lateinit var binding: ActivityLoginWebviewBinding
