+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/viewmodel/auth/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.takseha.presentation.viewmodel
+package com.takseha.presentation.viewmodel.auth
 
 import android.util.Log
 import androidx.lifecycle.LiveData
