+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/viewmodel/auth/PopupAgreementViewModel.kt
@@ -1,4 +1,4 @@
-package com.takseha.presentation.viewmodel
+package com.takseha.presentation.viewmodel.auth
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
