+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/viewmodel/auth/RegisterViewModel.kt
@@ -1,4 +1,4 @@
-package com.takseha.presentation.viewmodel
+package com.takseha.presentation.viewmodel.auth
 
 import android.app.Application
 import android.util.Log
