+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/viewmodel/home/MainHomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.takseha.presentation.viewmodel
+package com.takseha.presentation.viewmodel.home
 
 import android.app.Application
 import android.util.Log
